MyDicom improvement and decomposition
