+++ FrontEnd/src/components/CommonComponents/RessourcesDisplay/ReactTable/TableMyDicomPatientsStudies.js
@@ -1,37 +1,10 @@
-import React from 'react'
-import { useTable,useFilters,useRowSelect} from 'react-table'
+import React,{Component} from 'react'
 import ActionBouton from '../ActionBouton'
-import BTable from 'react-bootstrap/Table'
-import ColumnFilter from './ColumnFilter'
+import Table from './CommonSelectingAndFilteringTable'
 
-function App({tableData,onRowClick,onSelect}) {
-  
-  const Checkbox = React.forwardRef(
-    ({ indeterminate, ...rest }, ref) => {
-      const defaultRef = React.useRef()
-      const resolvedRef = ref || defaultRef
-  
-      React.useEffect(() => {
-        resolvedRef.current.indeterminate = indeterminate
-      }, [resolvedRef, indeterminate])
-  
-      return (
-        <>
-          <input type="checkbox" ref={resolvedRef} {...rest} />
-        </>
-      )
-    }
-  )
+export default class TableMyDicomPatientsStudies extends Component{
 
-  const defaultColumn = React.useMemo(
-    () => ({
-      Filter: ColumnFilter,
-    }),
-    []
-  )
-
-  const columns = React.useMemo(
-    () => [
+  columns = [
       {
         Header: 'Study Orthanc ID',
         accessor : 'StudyOrthancID',
@@ -82,86 +55,17 @@ function App({tableData,onRowClick,onSelect}) {
           )
         }
       },
-    ],
-    []
-  )
+    ]
 
-
-  const {
-    getTableProps,
-    getTableBodyProps,
-    headerGroups,
-    rows,
-    prepareRow,
-    selectedFlatRows
-  } = useTable(
-    {
-      columns,
-      data:tableData,
-      defaultColumn,
-      initialState: {
-        hiddenColumns: columns.map(column => {
-            if (column.hidden === true) return column.accessor || column.id;
-        })
-    },},
-    useFilters,
-    useRowSelect,
-    hooks => {
-      hooks.visibleColumns.push(columns => [
-        {
-          id: 'selection',
-          Header: ({ getToggleAllRowsSelectedProps }) => (
-            <div>
-              <Checkbox {...getToggleAllRowsSelectedProps()} />
-            </div>
-          ),
-          Cell: ({row}) => (
-            <div>
-              <Checkbox {...row.getToggleRowSelectedProps()} />
-            </div>
-          ),
-        },
-        ...columns,
-      ])
-    }
-  )
-
-  React.useEffect(() => { onSelect(selectedFlatRows); }, [selectedFlatRows.length]);
-  
-  return (
-    <>
-      <BTable striped bordered responsive {...getTableProps()}>
-        <thead>
-          {headerGroups.map(headerGroup => (
-            <tr {...headerGroup.getHeaderGroupProps()}>
-              {headerGroup.headers.map(column => (
-                <th {...column.getHeaderProps()}>{column.render('Header')}
-                 <div>{column.canFilter ? column.render('Filter') : null}</div>
-                </th>
-              ))}
-            </tr>
-          ))}
-        </thead>
-        <tbody {...getTableBodyProps()}>
-          {rows.map((row) => {
-            prepareRow(row)
-            return (
-              <React.Fragment key={row.getRowProps().key}>
-                <tr onClick={()=>onRowClick(row.values)}>
-                  {row.cells.map(cell => {
-                    return (
-                      <td {...cell.getCellProps()}>{cell.render('Cell')}</td>
-                    )
-                  })}
-                </tr>
-              </React.Fragment>
-            )
-          })}
-        </tbody>
-      </BTable>
-      <br/>
-    </>
-  )
+    render = () => {
+      return (
+          <Table 
+            tableData={this.props.data}
+            columns={this.columns}
+            onRowClick={this.props.onRowClick}
+            onSelect={this.props.onSelect}
+            rowStyle={this.props.rowStyle}
+          />
+      )
+  }
 }
-
-export default App
